perf(2dfea): lazy-load Zod import-path chunk to keep eager bundle small
Phase 9 of save/load JSON (plan §7 Phase 9, plan goal #18):

- Extract heavy import-side validation chain (Zod schema, semantic
  validator, migrations, applyToStore) into src/io/importPath.ts.
- promptUserForImport now does `await import('./importPath')` after
  the file picker resolves — runtime cost is one async chunk fetch on
  the first Import click, dwarfed by the file-picker dialog roundtrip.
- Export path (Ctrl+S, Toolbar Export button) stays eager so keystroke
  latency is unaffected.
- Move V1_UNITS const from schema.ts to schemaVersion.ts (with re-export
  for compatibility) so canonicalize.ts can read the unit policy without
  pulling Zod into the eager bundle.

Bundle-size verification:
  baseline:                180.85 KB gzipped (main chunk)
  before lazy-import:      198.39 KB gzipped (delta +17.54 KB — over budget)
  after lazy-import:       184.25 KB gzipped (delta +3.40 KB — within budget)
  lazy importPath chunk:    14.70 KB gzipped (loaded on first Import click)

Goal #18 (≤ 15 KB eager delta) met.
