+++ 2dfea/src/io/exportImport.ts
@@ -22,19 +22,12 @@
 import { useModelStore } from '../store/useModelStore';
 import { canonicalStringify } from './canonicalStringify';
 import { modelStateToFile } from './canonicalize';
-import { applyToStore } from './applyToStore';
-import {
-  ModelFileV1Schema,
-  warnUnknownKeys,
-  type ModelFileV1,
-} from './schema';
-import { semanticValidate } from './semanticValidator';
-import {
-  migrateToCurrent,
-  type AnyVersionedModel,
-} from './migrations';
-import { SchemaVersionError } from './schemaVersion';
 import { showToast } from '../store/useToastStore';
+
+// Heavy import-side modules are dynamic-imported from `promptUserForImport`
+// (plan §7 Phase 9 — bundle-size guard). See `./importPath.ts`. Tests
+// import `handleImportText` directly via the static re-export below;
+// runtime UI goes through the lazy code-split.
 
 // ----------------------------------------------------------------------------
 // Export
@@ -119,90 +112,10 @@ function pickFile(): Promise<string | null> {
   });
 }
 
-/**
- * Validate, migrate, and apply a parsed JSON text to the store. Pure
- * "given a string, do the import or surface the error". Public so the
- * round-trip helpers and tests can call it without the file-picker UI.
- */
-export async function handleImportText(text: string): Promise<void> {
-  // 1. JSON.parse
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(text);
-  } catch (err) {
-    showToast({
-      kind: 'error',
-      message: `Could not parse file: ${(err as Error).message}`,
-    });
-    return;
-  }
-
-  // 2. schemaVersion is a string
-  const version = (parsed as { schemaVersion?: unknown })?.schemaVersion;
-  if (typeof version !== 'string') {
-    showToast({
-      kind: 'error',
-      message: 'File is missing schemaVersion or it is not a string.',
-    });
-    return;
-  }
-
-  // 3. migrate
-  let migrated: AnyVersionedModel;
-  try {
-    migrated = migrateToCurrent(parsed as AnyVersionedModel);
-  } catch (err) {
-    if (err instanceof SchemaVersionError) {
-      showToast({
-        kind: 'error',
-        message: `Unsupported schema version "${err.receivedVersion}". This file was created by a newer version of 2dfea — please update.`,
-      });
-    } else {
-      showToast({
-        kind: 'error',
-        message: `Migration failed: ${(err as Error).message}`,
-      });
-    }
-    return;
-  }
-
-  // 4. Zod parse
-  const result = ModelFileV1Schema.safeParse(migrated);
-  if (!result.success) {
-    const issues = result.error.issues;
-    const first = issues[0];
-    const path = first.path.join('.');
-    showToast({
-      kind: 'error',
-      message: `Invalid model file: ${path ? path + ' — ' : ''}${first.message}`,
-    });
-    console.warn('[Import] Zod validation errors:', issues);
-    return;
-  }
-
-  // 5. semantic validate
-  const semantic = semanticValidate(result.data.model);
-  if (semantic.length > 0) {
-    showToast({
-      kind: 'error',
-      message: `Model is structurally invalid: ${semantic[0]}${
-        semantic.length > 1 ? ` (and ${semantic.length - 1} more — see console)` : ''
-      }`,
-    });
-    console.warn('[Import] Semantic validation errors:', semantic);
-    return;
-  }
-
-  // 6. emit unknown-keys diagnostic (single warn, deduped) before mutating
-  warnUnknownKeys(result.data as ModelFileV1);
-
-  // 7. apply
-  applyToStore(result.data as ModelFileV1);
-  console.log(
-    `[Import] Validation succeeded; applied ${result.data.model.nodes.length} nodes, ${result.data.model.elements.length} elements, ${result.data.model.loads.nodal.length + result.data.model.loads.distributed.length + result.data.model.loads.elementPoint.length} loads.`
-  );
-  showToast({ kind: 'info', message: 'Model imported successfully.' });
-}
+// handleImportText lives in `./importPath` and is dynamic-imported from
+// `promptUserForImport` to keep Zod + semantic validator + applyToStore
+// off the eager bundle. Tests and other non-UI consumers should import
+// directly from `./importPath`.
 
 /**
  * Top-level "user clicked Import" flow. Two confirm gates fire in order:
@@ -239,5 +152,10 @@ export async function promptUserForImport(): Promise<void> {
 
   const text = await pickFile();
   if (text === null) return; // user cancelled or read failed
+  // Lazy-load the validator + apply chain (Zod, semantic validator,
+  // applyToStore). Runtime cost: a single async chunk fetch on the
+  // first import click; insignificant compared to the file picker
+  // round-trip.
+  const { handleImportText } = await import('./importPath');
   await handleImportText(text);
 }
